Added protection around receiving unknown packets
Removed some 32-bit incrementers from NetGetData as they might be
causing the overflow exceptions.  As this is non-reproducible at
this point, cannot mark is as fixed yet.

Possibly fixes NyxStudios/TShock#1077

+++ Terraria/MessageBuffer.cs
@@ -72,7 +72,7 @@ namespace Terraria
 			num2 = start + 1;
 			num1 = this.readBuffer[start];
 
-			if (Enum.IsDefined(typeof(PacketTypes), num1) == false)
+			if (Enum.IsDefined(typeof(PacketTypes), (int)num1) == false)
 			{
 				return;
 			}
